@@ -2,7 +2,7 @@
 name = "stock-check-backend"
 version = "0.1.0"
 description = "Backend for Stock Check application with AI Agent"
-requires-python = ">=3.9"
+requires-python = ">=3.11"
 dependencies = [
     # Web framework
     "fastapi>=0.109.2",
@@ -34,6 +34,9 @@ dependencies = [
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["app"]
 
 [tool.uv]
 dev-dependencies = [
